splitTagsを別ファイルに分離
