+++ src/app/art/[art_id]/edit/tag/page.tsx
@@ -2,8 +2,7 @@ import { getArt } from "@/art/get"
 import { EditArtTagForm } from "./EditArtTagForm"
 import { notFound } from "next/navigation"
 import { getTags } from "@/art/tag/getTags"
-import { getGenreTags, medias, others } from "@/art/components/tag/tags"
-import { ArtTag } from "@/art/type"
+import { splitTags } from "@/art/tag/split"
 
 interface PageProps {
     params: { art_id: string }
@@ -31,35 +30,3 @@ const ArtEditTagPage = async ({ params: { art_id } }: PageProps) => {
 }
 export default ArtEditTagPage
 
-const splitTags = (tags: ArtTag[]) => {
-    const mediaTags: ArtTag[] = []
-    const genreTags: ArtTag[] = []
-    const otherTags: ArtTag[] = []
-    // media と other
-    const notGroupedTags: ArtTag[] = []
-    tags.forEach((tag) => {
-        if (medias.includes(tag)) {
-            mediaTags.push(tag)
-        } else if (others.includes(tag)) {
-            otherTags.push(tag)
-        } else {
-            notGroupedTags.push(tag)
-        }
-    })
-    // genre
-    const genres = getGenreTags(mediaTags)
-    notGroupedTags.forEach((tag, index) => {
-        if (genres.includes(tag)) {
-            genreTags.push(tag)
-            notGroupedTags.splice(index, 1)
-        }
-    })
-    // originalTags
-    const originalTags: ArtTag[] = notGroupedTags
-    return {
-        mediaTags,
-        genreTags,
-        otherTags,
-        originalTags,
-    }
-}
